Update phpunit 10 - part 3.
